refactor: remove legacy prompt system, unify on skills
Remove the entire old prompt stages system that has been superseded by
the unified skill system. All prompt-based functionality now goes through
the skills collection and skill APIs.

Deleted files:
- IPromptService, PromptService (promptstages CRUD)
- PromptStagesController (admin + client endpoints)
- PromptStagesOptimizeController (AI prompt optimization)
- PromptSettings/PromptEntry/RolePrompt/SkillSettings/SkillEntry models
- PromptStageRequests/PromptStageOptimizeRequests DTOs
- PromptStageOptimizeStreamEvent
- Old SkillsController (/api/v1/skills reading skill_settings)
- Desktop get_prompts Tauri command + PromptClientItem types
- Admin PromptStagesPage + prompts service/contracts

Modified:
- SkillService: removed IPromptService dependency and migration code
- ISkillService: removed MigrateFromPromptsAsync
- PrdAgentController: removed IPromptService, old GetPrompts endpoint
- Program.cs: removed IPromptService DI and startup migration
- MongoDbContext: removed Prompts/PromptsRaw/SkillSettings collections
- BsonClassMapRegistration: removed RegisterPromptSettings
- Skill.cs: moved SkillParameter class here from deleted PromptSettings.cs
- Desktop sessionStore: removed prompts/promptsUpdatedAt fields, bumped version
- Admin api.ts/index.ts/authzMenuMapping: removed old prompt entries

Kept intact:
- ISystemPromptService (role-based base system prompts, still used by ChatService)
- SystemPromptsController (admin management of system prompts)
- PromptOverrides (image gen plan overrides, separate concern)
